requirements.txt
load_dotenv
langchain
langchain-google-genai
langchain-community
pypdf
google-generativeai
chromadb